Changed create account to go to login page
When the user creates an account they are redirected to the login page instead of logging in incorrectly.

Also made it so the user can hit enter on the password field on the login page to login, and on the confirm password field on the signup page to create an account.